Refactor server endpoints
