--- server/main.go
@@ -1,9 +1,0 @@
-package main
-
-func main() {
-
-	app := App{}
-	app.Init()
-	app.Run(":8080")
-
-}
